feat/#1: 컴포넌트 구조 재설계
