+++ app/auth/page.tsx
@@ -5,11 +5,10 @@ import RequestAccessToken from '@/components/auth/RequestAccessToken';
 
 const page = () => {
   return (
-    <RequestAccessToken
+    <LoginLanding
       clientId={process.env.GOOGLE_CLIENT_ID as string}
-      clientSecret={process.env.GOOGLE_CLIENT_SECRET as string}>
-      <LoginLanding />
-    </RequestAccessToken>
+      clientSecret={process.env.GOOGLE_CLIENT_SECRET as string}
+    />
   );
 };
